+++ tests/scalers/kafka.test.ts
@@ -7,6 +7,7 @@ const defaultNamespace = 'kafka-test'
 const defaultCluster = 'kafka-cluster'
 const timeToWait = 300
 const defaultTopic = 'kafka-topic'
+const defaultTopic2 = 'kafka-topic-2'
 const defaultKafkaClient = 'kafka-client'
 const strimziOperatorVersion = '0.18.0'
 const commandToCheckReplicas = `kubectl get deployments/kafka-consumer --namespace ${defaultNamespace} -o jsonpath="{.spec.replicas}"`
@@ -17,11 +18,13 @@ const kafkaTopicYamlFile = tmp.fileSync()
 const kafkaClientYamlFile = tmp.fileSync()
 const kafkaApplicationLatestYamlFile = tmp.fileSync()
 const kafkaApplicationEarliestYamlFile = tmp.fileSync()
+const kafkaApplicationMultipleTopicsYamlFile = tmp.fileSync()
 const scaledObjectEarliestYamlFile = tmp.fileSync()
 const scaledObjectLatestYamlFile = tmp.fileSync()
+const scaledObjectMultipleTopicsYamlFile = tmp.fileSync()
 
 test.before('Set up, create necessary resources.', t => {
-	sh.config.silent = true
+    sh.config.silent = true
 	sh.exec(`kubectl create namespace ${defaultNamespace}`)
 
   const strimziOperatorYaml = sh.exec(`curl -L https://github.com/strimzi/strimzi-kafka-operator/releases/download/${strimziOperatorVersion}/strimzi-cluster-operator-${strimziOperatorVersion}.yaml`).stdout
@@ -42,9 +45,9 @@ test.before('Set up, create necessary resources.', t => {
 		0,
 		sh.exec(`kubectl wait kafka/${defaultCluster} --for=condition=Ready --timeout=${timeToWait}s --namespace ${defaultNamespace}`).code,
 		'Kafka instance should be ready within given time limit.'
-  )
+    )
 
-	fs.writeFileSync(kafkaTopicYamlFile.name, kafkaTopicYaml)
+    fs.writeFileSync(kafkaTopicYamlFile.name, kafkaTopicsYaml)
 	t.is(
 		0,
 		sh.exec(`kubectl apply -f ${kafkaTopicYamlFile.name} --namespace ${defaultNamespace}`).code,
@@ -53,8 +56,13 @@ test.before('Set up, create necessary resources.', t => {
 	t.is(
 		0,
 		sh.exec(`kubectl wait kafkatopic/${defaultTopic} --for=condition=Ready --timeout=${timeToWait}s --namespace ${defaultNamespace}`).code,
-		'Kafka topic should be ready within given time limit.'
-  )
+		'Kafka topic should be ready withlanguage-mattersin given time limit.'
+    )
+    t.is(
+        0,
+        sh.exec(`kubectl wait kafkatopic/${defaultTopic2} --for=condition=Ready --timeout=${timeToWait}s --namespace ${defaultNamespace}`).code,
+        'Kafka topic2 should be ready within given time limit.'
+    )
 
 	fs.writeFileSync(kafkaClientYamlFile.name, kafkaClientYaml)
 	t.is(
@@ -66,7 +74,7 @@ test.before('Set up, create necessary resources.', t => {
 		0,
 		sh.exec(`kubectl wait pod/${defaultKafkaClient} --for=condition=Ready --timeout=${timeToWait}s --namespace ${defaultNamespace}`).code,
 		'Kafka client should be ready within given time limit.'
-  )
+    )
 
   fs.writeFileSync(kafkaApplicationEarliestYamlFile.name, kafkaApplicationEarliestYaml)
 
@@ -170,7 +178,6 @@ test.serial('Applying ScaledObject latest policy should not scale up pods', t =>
   sh.exec(`sleep 5s`)
   waitForReplicaCount(1, commandToCheckReplicas)
   t.is('0', sh.exec(commandToCheckReplicas).stdout, 'Replica count should be 0.')
-
 })
 
 
@@ -187,11 +194,80 @@ test.serial('Latest Scale object should scale with new messages', t => {
   }
 })
 
+test.serial('Cleanup after latest policy test', t=> {
+    t.is(
+        0,
+        sh.exec(`kubectl delete -f ${scaledObjectLatestYamlFile.name} --namespace ${defaultNamespace}`).code,
+        'Deleting Scaled Object should work.'
+    )
+    t.is(
+        0,
+        sh.exec(`kubectl delete -f ${kafkaApplicationLatestYamlFile.name} --namespace ${defaultNamespace}`).code,
+        'Deleting kafka application should work.'
+    )
+    sh.exec(`sleep 10s`)
+})
+
+test.serial('Applying ScaledObject with multiple topics should scale up pods', t => {
+    // Make the consumer commit the all offsets for all topics in the group
+    sh.exec(`kubectl exec --namespace ${defaultNamespace} ${defaultKafkaClient} -- sh -c 'kafka-console-consumer --bootstrap-server "${defaultCluster}-kafka-bootstrap.${defaultNamespace}:9092" --topic ${defaultTopic}  --group multiTopic --from-beginning --consumer-property enable.auto.commit=true --timeout-ms 15000'`)
+    sh.exec(`kubectl exec --namespace ${defaultNamespace} ${defaultKafkaClient} -- sh -c 'kafka-console-consumer --bootstrap-server "${defaultCluster}-kafka-bootstrap.${defaultNamespace}:9092" --topic ${defaultTopic2} --group multiTopic --from-beginning --consumer-property enable.auto.commit=true --timeout-ms 15000'`)
+
+    fs.writeFileSync(kafkaApplicationMultipleTopicsYamlFile.name, kafkaApplicationMultipleTopicsYaml)
+    t.is(
+        0,
+        sh.exec(`kubectl apply -f ${kafkaApplicationMultipleTopicsYamlFile.name} --namespace ${defaultNamespace}`).code,
+        'Deploying Kafka application should work.'
+    )
+    sh.exec(`sleep 5s`)
+    fs.writeFileSync(scaledObjectMultipleTopicsYamlFile.name, scaledObjectMultipleTopicsYaml)
+
+    t.is(
+        0,
+        sh.exec(`kubectl apply -f ${scaledObjectMultipleTopicsYamlFile.name} --namespace ${defaultNamespace}`).code,
+        'Deploying Scaled Object should work.'
+    )
+    sh.exec(`sleep 5s`)
+    waitForReplicaCount(1, commandToCheckReplicas)
+    t.is('0', sh.exec(commandToCheckReplicas).stdout, 'Replica count should be 0.')
+
+    // produce a single msg to the default topic should not scale
+    sh.exec(`kubectl exec --namespace ${defaultNamespace} ${defaultKafkaClient} -- sh -exc 'echo "{\"text\": \"foo\"}" | kafka-console-producer --broker-list ${defaultCluster}-kafka-bootstrap.${defaultNamespace}:9092 --topic ${defaultTopic}'`)
+    sh.exec(`sleep 20s`)
+    waitForReplicaCount(0, commandToCheckReplicas)
+    t.is('0', sh.exec(commandToCheckReplicas).stdout, 'Replica count should be 0.')
+
+    // produce one more msg to the different topic should trigger scale
+    sh.exec(`kubectl exec --namespace ${defaultNamespace} ${defaultKafkaClient} -- sh -exc 'echo "{\"text\": \"foo\"}" | kafka-console-producer --broker-list ${defaultCluster}-kafka-bootstrap.${defaultNamespace}:9092 --topic ${defaultTopic2}'`)
+    sh.exec(`sleep 20s`)
+    waitForReplicaCount(1, commandToCheckReplicas)
+    t.is('1', sh.exec(commandToCheckReplicas).stdout, 'Replica count should be 1.')
+})
+
+test.serial('Cleanup after multiple topics test', t=> {
+    t.is(
+        0,
+        sh.exec(`kubectl delete -f ${scaledObjectMultipleTopicsYamlFile.name} --namespace ${defaultNamespace}`).code,
+        'Deleting Scaled Object should work.'
+    )
+    t.is(
+        0,
+        sh.exec(`kubectl delete -f ${kafkaApplicationMultipleTopicsYamlFile.name} --namespace ${defaultNamespace}`).code,
+        'Deleting kafka application should work.'
+    )
+})
+
+
 test.after.always('Clean up, delete created resources.', t => {
   const resources = [
     `${scaledObjectEarliestYamlFile.name}`,
     `${scaledObjectLatestYamlFile.name}`,
+    `${scaledObjectMultipleTopicsYamlFile.name}`,
+
+    `${kafkaApplicationEarliestYamlFile.name}`,
     `${kafkaApplicationLatestYamlFile.name}`,
+    `${kafkaApplicationMultipleTopicsYamlFile.name}`,
+
     `${kafkaClientYamlFile.name}`,
     `${kafkaTopicYamlFile.name}`,
     `${kafkaClusterYamlFile.name}`,
@@ -231,10 +307,24 @@ spec:
     topicOperator: {}
     userOperator: {}`
 
-const kafkaTopicYaml = `apiVersion: kafka.strimzi.io/v1beta1
+const kafkaTopicsYaml = `apiVersion: kafka.strimzi.io/v1beta1
 kind: KafkaTopic
 metadata:
   name: ${defaultTopic}
+  labels:
+    strimzi.io/cluster: ${defaultCluster}
+  namespace: ${defaultNamespace}
+spec:
+  partitions: 3
+  replicas: 1
+  config:
+    retention.ms: 604800000
+    segment.bytes: 1073741824
+---
+apiVersion: kafka.strimzi.io/v1beta1
+kind: KafkaTopic
+metadata:
+  name: ${defaultTopic2}
   labels:
     strimzi.io/cluster: ${defaultCluster}
   namespace: ${defaultNamespace}
@@ -310,6 +400,40 @@ spec:
           - -c
           - "kafka-console-consumer --bootstrap-server ${defaultCluster}-kafka-bootstrap.${defaultNamespace}:9092 --topic ${defaultTopic} --group earliest --from-beginning --consumer-property enable.auto.commit=false"`
 
+const kafkaApplicationMultipleTopicsYaml = `
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: kafka-consumer
+  namespace: ${defaultNamespace}
+  labels:
+    app: kafka-consumer
+spec:
+  selector:
+    matchLabels:
+      app: kafka-consumer
+  template:
+    metadata:
+      labels:
+        app: kafka-consumer
+    spec:
+      containers:
+      # only recent version of kafka-console-consumer support flag "include"
+      # old version's equiv flag will violate language-matters commit hook
+      # work around -> create two consumer container joining the same group
+      - name: kafka-consumer
+        image: confluentinc/cp-kafka:5.2.1
+        command:
+          - sh
+          - -c
+          - "kafka-console-consumer --bootstrap-server ${defaultCluster}-kafka-bootstrap.${defaultNamespace}:9092 --topic '${defaultTopic}'  --group multiTopic --from-beginning --consumer-property enable.auto.commit=false"
+      - name: kafka-consumer-2
+        image: confluentinc/cp-kafka:5.2.1
+        command:
+          - sh
+          - -c
+          - "kafka-console-consumer --bootstrap-server ${defaultCluster}-kafka-bootstrap.${defaultNamespace}:9092 --topic '${defaultTopic2}' --group multiTopic --from-beginning --consumer-property enable.auto.commit=false"`
+
 const scaledObjectEarliestYaml = `apiVersion: keda.sh/v1alpha1
 kind: ScaledObject
 metadata:
@@ -342,4 +466,20 @@ spec:
       bootstrapServers: ${defaultCluster}-kafka-bootstrap.${defaultNamespace}:9092
       consumerGroup: latest
       lagThreshold: '1'
+      offsetResetPolicy: 'latest'`
+
+const scaledObjectMultipleTopicsYaml = `apiVersion: keda.sh/v1alpha1
+kind: ScaledObject
+metadata:
+  name: kafka-consumer-multi-topic
+  namespace: ${defaultNamespace}
+spec:
+  scaleTargetRef:
+    name: kafka-consumer
+  triggers:
+  - type: kafka
+    metadata:
+      bootstrapServers: ${defaultCluster}-kafka-bootstrap.${defaultNamespace}:9092
+      consumerGroup: multiTopic
+      lagThreshold: '2'
       offsetResetPolicy: 'latest'`